+++ tikkle-be/src/main/java/DabuOps/tikkle/transaction_history/controller/TransactionHistoryController.java
@@ -7,8 +7,6 @@ import DabuOps.tikkle.transaction_history.entity.TransactionHistory;
 import DabuOps.tikkle.transaction_history.mapper.TransactionHistoryMapper;
 import DabuOps.tikkle.transaction_history.repository.TransactionHistoryRepository;
 import DabuOps.tikkle.transaction_history.service.TransactionHistoryService;
-import DabuOps.tikkle.userauth.dto.AccountTransactionDto;
-import DabuOps.tikkle.userauth.mapper.AccountTransactionMapper;
 import DabuOps.tikkle.userauth.service.UserAuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -32,8 +30,6 @@ public class TransactionHistoryController {
     private final TransactionHistoryMapper mapper;
 
     private final UserAuthService userAuthService;
-    private final AccountTransactionMapper accountTransactionMapper;
-
     @PostMapping()
     public ResponseEntity postTransactionHistory(@Valid @RequestBody TransactionHistoryDto.Post requestBody) {
         Long memberCategoryId = requestBody.getMemberCategoryId();
